udev: fibre channel: add support to NPIV FC virtual interfaces.
When using NPIV, you can create multiple virtual HBAs on top of the
physical one, this means that the physical N_Port can have multiple
port IDs associated to it.  Suppose a device is assigned to the
physical HBA and a virtual HBA, the original code uses only the port
ID of the physical N_Port to build the by-path link and udev ends up
trying to create two by-path links with the same name.

This patch fixes the problem by using the port ID of the virtual HBA
whenever it detects that the device belongs to a virtual HBA,
otherwise it uses the port ID of the physical HBA.

(cherry picked from commit 155a760bcedd11b7f3b430350a46f10736286895)

Resolves: #1032218